pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "training-signal-processing"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.5.2",
]

[tool.uv]
package = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[dependency-groups]
dev = [
    "import-linter>=2.11,<3",
    "pytest>=9.0.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.15.11",
]
model = [
    "torch>=2.8,<2.9",
    "transformers>=4.51.0",
]
remote_ocr = [
    "boto3>=1.39.0",
    "click>=8.1.8",
    "marker-pdf>=1.10.0",
    "mlflow-skinny>=2.22.0",
    "pyarrow>=20.0.0",
    "pyyaml>=6.0.2",
    "ray[data]>=2.49.0",
    "torch>=2.8,<2.9",
    "tqdm>=4.67.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.importlinter]
root_packages = [
    "training_signal_processing.core",
    "training_signal_processing.ops",
    "training_signal_processing.pipelines",
]

[[tool.importlinter.contracts]]
name = "shared layers must not import from pipelines"
type = "forbidden"
source_modules = [
    "training_signal_processing.core",
    "training_signal_processing.ops",
]
forbidden_modules = [
    "training_signal_processing.pipelines",
]
as_packages = true

[tool.hatch.build.targets.wheel]
packages = ["src/training_signal_processing"]
